+++ dbt_platform_helper/commands/notify.py
@@ -2,7 +2,7 @@ import click
 from slack_sdk import WebClient
 from slack_sdk.models import blocks
 
-from dbt_platform_helper.providers.platform_helper_version import (
+from dbt_platform_helper.providers.platform_helper_versioning import (
     PlatformHelperVersioning,
 )
 from dbt_platform_helper.utils.arn_parser import ARN
